Update travis script.

+++ .travis/travis.sh
@@ -4,22 +4,28 @@ EFXPHP_PATH=$(pwd)
 
 # PHP Settings
 PHP_NAME="php-$PHP_VERSION"
-PHP_PACKAGE="php-$PHP_VERSION-travis.tar.gz"
+PHP_PACKAGE="php-$PHP_VERSION.tar.gz"
 PHP_URL="http://us1.php.net/get/$PHP_PACKAGE/from/this/mirror"
 PHP_CLI="/home/travis/build/emilkm/efxphp/php-7.1.1/install/bin/php"
 
 # Move out of project
 BASE_PATH=$(pwd)
 
- # Get and extract prebuilt PHP
-wget "$PHP_URL" -O $PHP_PACKAGE
-tar -xf $PHP_PACKAGE
 
-# Build PHP
-./buildconf --force
-./configure --enable-debug --disable-all --enable-libxml --enable-simplexml --enable-dom --with-phar --prefix=/home/travis/build/emilkm/efxphp/php-7.1.1/install
-make install
+if [ -e "/home/travis/build/emilkm/efxphp/php-7.1.1/configure" ] 
+then
+	echo "File /home/travis/build/emilkm/efxphp/php-7.1.1/configure exists." 
+else
+    # Get and extract prebuilt PHP
+	wget "$PHP_URL" -O $PHP_PACKAGE
+	tar -xf $PHP_PACKAGE
 
+	# Build PHP
+	cd $PHP_NAME
+	./buildconf --force
+	./configure --enable-debug --disable-all --enable-libxml --enable-simplexml --enable-dom --with-phar --prefix=/home/travis/build/emilkm/efxphp/php-7.1.1/install
+	make install
+fi
 
 # Get AMFEXT
 git clone --depth 1 https://github.com/emilkm/amfext.git amfext
